+++ README.md
@@ -26,10 +26,10 @@ This extension requires that Node.js and npm be installed on the build agent. Th
 | dockerImageName           | The name of the container image to test.                                                                                                                                                                            | yes, if container image test | none          | string                                                                            |
 | dockerfilePath            | The path to the Dockerfile corresponding to the `dockerImageName`                                                                                                                                                   | yes, if container image test | none          | string                                                                            |
 | targetFile                | Applicable to application type tests ony. The path to the manifest file to be used by Snyk. Should only be provided if non-standard.                                                                                | no                           | none          | string                                                                            |
-| severityThreshold         | The severity-threshold to use when testing. By default, issues of all severity types will be found.                                                                                                                 | no                           | "low"         | string: "low" or "medium" or "high" or "critical"                                 |
-| failOnThreshold           | The severity-threshold to use when determining if the build should fail. Combine with `failOnIssues` to get fine grained control over reporting and build failure behaviour. For example, with `failOnIssues` set to true and `failOnThreshold` to `critical`, all issues would be reported on but only critical issues would cause a build failure                                                                                                            | no                           | "low"         | string: "low" or "medium" or "high" or "critical"                                 |
+| severityThreshold         | The severity-threshold to use when testing and reporting. By default, issues of all severity types will be found.                                                                                                                 | no                           | "low"         | string: "low" or "medium" or "high" or "critical"                                 |
+| failOnThreshold           | The `severityThreshold` parameter is used to control the interaction with the Snyk CLI and reporting vulnerabilities. The `failOnThreshold` gives you additional control over build failure behaviour. For example, with `failOnIssues` set to `true` and `failOnThreshold` to `critical`, all issues would be reported on but _only_ critical issues would cause a build failure. See Usage Examples for more information                                                                                                            | no                           | "low"         | string: "low" or "medium" or "high" or "critical"                                 |
 | monitorWhen               | When to run `snyk monitor`. Valid options are `always` (default), `noIssuesFound`, and `never`. If set, this option overrides the value of `monitorOnBuild`.                                                        | no                           | "always"      | boolean                                                                           |
-| failOnIssues              | This specifies if builds should be failed or continued based on issues found by Snyk.                                                                                                                               | yes                          | true          | boolean                                                                           |
+| failOnIssues              | This specifies if builds should be failed or continued based on issues found by Snyk. Combine with `failOnThreshold` to control which severity of issues causes the build to fail                                                                                                                               | yes                          | true          | boolean                                                                           |
 | projectName               | A custom name for the Snyk project to be created on snyk.io                                                                                                                                                         | no                           | none          | string                                                                            |
 | organization              | Name of the Snyk organisation name, under which this project should be tested and monitored                                                                                                                         | no                           | none          | string                                                                            |
 | testDirectory             | Alternate working directory. For example, if you want to test a manifest file in a directory other than the root of your repo, you would put in relative path to that directory.                                    | no                           | none          | string                                                                            |
@@ -67,13 +67,14 @@ Here's a full example:
     monitorWhen: 'always'
 ```
 
-An example that uses the default value for `severityThreshold` (low) but configures `failOnThreshold` to critical. This configuration would _only fail_ the build when critical issues are found, but all issues would be reported back to your snyk project for analysis
+An example that specifies a value for `severityThreshold` as medium  and configures `failOnThreshold` to critical. This configuration would _only fail_ the build when critical issues are found, but all issues detected at medium, high and critical would be reported back to your snyk project for analysis
 
 ```
 - task: SnykSecurityScan@1
   inputs:
     serviceConnectionEndpoint: 'mySnykToken'
     testType: 'app'
+    severityThreshold: 'medium'    
     failOnIssues: true
     failOnThreshold: 'critical'
     monitorWhen: 'always'
